Use sveltekit-adapter-chrome-extension adapter and update build
