--- chrome/fonts
@@ -1,1 +1,0 @@
-../build/fonts
