--- chrome/images
@@ -1,1 +1,0 @@
-../build/images
