--- chrome/index.html
@@ -1,1 +1,0 @@
-../build/index.html
