+++ package.json
@@ -4,11 +4,10 @@
   "description": "A countdown with a fresh wallpapered tab on every new tab",
   "scripts": {
     "dev": "vite dev",
-    "build": "yarn build:app && yarn build:extension",
+    "build": "yarn build:app && yarn build:background-js && yarn build:version-manifest",
     "build:app": "vite build",
-    "build:extension": "yarn build:manifest && yarn build:background-js",
-    "build:manifest": "node scripts/build-manifest.js",
     "build:background-js": "rollup -c scripts/background-js-rollup.config.js",
+    "build:version-manifest": "node scripts/version-manifest.js",
     "package": "yarn _package && echo \"📦  Package created @ build/vorfreude.zip\"",
     "preview": "vite preview",
     "test": "yarn test:unit && yarn test:acceptance",
@@ -58,6 +57,7 @@
     "svelte-inline-compile": "^0.2.0",
     "svelte-jester": "^2.3.2",
     "svelte-preprocess": "^4.10.7",
+    "sveltekit-adapter-chrome-extension": "^1.3.0",
     "ts-node": "^10.9.1",
     "tslib": "^2.4.0",
     "typescript": "~4.8.3",
