+++ static/settings.html
@@ -1,4 +1,3 @@
 <script>
   document.location = "index.html#settings";
 </script>
-
